requirements.txt
ccxt
pandas
numpy
ta
python-dotenv
scikit-learn
joblib
xgboost
